+++ spec/requests/messsage_addition_spec.rb
@@ -3,10 +3,10 @@ require 'spec_helper'
 describe 'A status request with a check that modifies messages' do
   subject { get(status_path) ; response }
 
-  context 'creating a message with a lambda' do
+  context 'creating a message with a block' do
     before do
       Rapporteur::Checker.clear
-      Rapporteur::Checker.add_check(Proc.new { |checker| checker.add_message('git_repo', 'git@github.com:organization/repo.git') })
+      Rapporteur::Checker.add_check { |checker| checker.add_message('git_repo', 'git@github.com:organization/repo.git') }
     end
 
     context 'with an unerring response' do
@@ -19,7 +19,7 @@ describe 'A status request with a check that modifies messages' do
 
     context 'with an erring response' do
       before do
-        Rapporteur::Checker.add_check(Proc.new { |checker| checker.add_error('failed') })
+        Rapporteur::Checker.add_check { |checker| checker.add_error('failed') }
       end
 
       it_behaves_like 'an erred status response'