+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 ## [HEAD][unreleased] / unreleased
 
-* No significant changes.
+* Removed active_model_serializers dependency.
+* Extracted time and revision checks into Checks::TimeCheck and
+  Checks::RevisionCheck, and applied them as the default checks.
+* Updated Checker#add_error to allow for I18n interpolated values.
+* Updated Checker#add_check to take a block in addition to a lambda or object
+  that responds to #call.
+
+### :boom: Backward incompatible changes
+
+* Flattened the messages key in the JSON response. All messages are now
+  included at the top level of the hash.
+* It's now possible to remove all checks by calling Checker#clear. This
+  includes the default TimeCheck and RevisionCheck checks.
 
 ## [1.1.0][v1.1.0] / 2013-05-30